refactoring things out of SliverAPI.py
