+++ Payload_Type/sliverimplant/sliverimplant/SliverRequests/SliverAPI.py
@@ -6,7 +6,7 @@ from mythic_container.MythicCommandBase import *
 from mythic_container.MythicRPC import *
 from mythic_container.PayloadBuilder import *
 import json
-import gzip
+
 from mythic_container.LoggingBase import *
 from mythic_container.MythicGoRPC import *
 
@@ -47,245 +47,6 @@ async def create_sliver_interact(taskData: PTTaskMessageAllData):
     }
 
     return interact, isBeacon
-
-async def generate():
-    # TODO: generate an implant based on config provided
-
-    # implant_config = client_pb2.ImplantConfig(
-    #     IsBeacon=True,
-    #     Name="sliver-pytest-1",
-    #     GOARCH="amd64",
-    #     GOOS="linux",
-    #     Format=client_pb2.OutputFormat.EXECUTABLE,
-    #     ObfuscateSymbols=False,
-    #     C2=[client_pb2.ImplantC2(Priority=0, URL="http://localhost:80")],
-    # )
-
-    # implant = await client.generate_implant(implant_config)
-
-    return None
-
-async def ifconfig(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    ifconfig_results = await interact.ifconfig()
-
-    if (isBeacon):
-        ifconfig_results = await ifconfig_results
-
-    return ifconfig_results
-
-async def info(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    #             Session ID: d7e28b37-88be-44f9-ba31-8913bf535d1a
-    #               Name: FUNNY_DRIVEWAY
-    #           Hostname: ubuntu
-    #               UUID: c744e366-d14c-4bf3-94c3-558012eda8a1
-    #           Username: root
-    #                UID: 0
-    #                GID: 0
-    #                PID: 120952
-    #                 OS: linux
-    #            Version: Linux ubuntu 6.5.0-27-generic
-    #             Locale: en-US
-    #               Arch: amd64
-    #          Active C2: mtls://192.168.17.129:443
-    #     Remote Address: 192.168.17.129:50072
-    #          Proxy URL: 
-    # Reconnect Interval: 1m0s
-    #      First Contact: Fri Apr 12 21:11:25 CDT 2024 (28m16s ago)
-    #       Last Checkin: Fri Apr 12 21:39:39 CDT 2024 (2s ago)
-
-    responseString = []
-    responseString.append(f"Session Id: {interact.session_id}")
-    responseString.append(f"Name: {interact.name}")
-    responseString.append(f"Hostname: {interact.hostname}")
-    responseString.append(f"UUID: {interact.uuid}")
-    responseString.append(f"Username: {interact.username}")
-    responseString.append(f"UID: {interact.uid}")
-    responseString.append(f"GID: {interact.gid}")
-    responseString.append(f"PID: {interact.pid}")
-    responseString.append(f"OS: {interact.os}")
-    responseString.append(f"Version: {interact.version}")
-    # responseString.append(f"Locale: {interact.L}")
-    responseString.append(f"Arch: {interact.arch}")
-    responseString.append(f"Active C2: {interact.active_c2}")
-    responseString.append(f"Remote Address: {interact.remote_address}")
-    responseString.append(f"Proxy URL: {interact.proxy_url}")
-    responseString.append(f"Reconnect Interval: {interact.reconnect_interval}")
-    # responseString.append(f"First Contact: {interact.last_checkin}")
-    responseString.append(f"Last Checkin: {interact.last_checkin}")
-
-    finalResponse = "\n".join(responseString)
-
-    return finalResponse
-
-async def ping(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-    ping_result = await interact.ping()
-
-    if (isBeacon):
-        ping_result = await ping_result
-
-    return f"{ping_result}"
-
-async def screenshot(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-    screenshot_result = await interact.screenshot()
-
-    if (isBeacon):
-        screenshot_result = await screenshot_result
-
-    # decompressed = gzip.decompress(screenshot_result.Data)
-
-    return screenshot_result.Data
-
-async def whoami(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    # TODO: this is special for windows
-
-    return f"Logon ID: {interact.username}"
-
-async def getgid(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-    return f"{interact.gid}"
-
-async def getpid(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-    return f"{interact.pid}"
-
-async def getuid(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-    return f"{interact.uid}"
-
-async def download(taskData: PTTaskMessageAllData, full_path: str):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    download_results = await interact.download(remote_path=full_path)
-
-    if (isBeacon):
-        download_results = await download_results
-
-    plaintext = gzip.decompress(download_results.Data)
-
-    return plaintext
-
-async def upload(taskData: PTTaskMessageAllData, agent_file_uuid: str, path: str):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    filestuff = await SendMythicRPCFileGetContent(MythicRPCFileGetContentMessage(
-        AgentFileId=agent_file_uuid
-    ))
-
-    upload_results = await interact.upload(
-        remote_path=path,
-        data=filestuff.Content
-    )
-
-    if (isBeacon):
-        upload_results = await upload_results     
-
-    return upload_results
-
-async def rm(taskData: PTTaskMessageAllData, path_to_rm: str):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    rm_results = await interact.rm(remote_path=path_to_rm)
-    
-    if (isBeacon):
-        rm_results = await rm_results
-
-    return rm_results
-
-async def ls(taskData: PTTaskMessageAllData, path_to_ls: str):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    ls_results = await interact.ls(remote_path=path_to_ls)
-
-    if (isBeacon):
-        ls_results = await ls_results
-        
-    return ls_results
-
-async def ps(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    ps_results = await interact.ps()
-
-    if (isBeacon):
-        ps_results = await ps_results
-
-    return ps_results
-
-async def netstat(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    netstat_results = await interact.netstat(tcp=True, udp=True, ipv4=True, ipv6=True, listening=True)
-
-    if (isBeacon):
-        netstat_results = await netstat_results
-
-    return netstat_results
-
-async def cd(taskData: PTTaskMessageAllData, remote_path: str):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    cd_results = await interact.cd(remote_path=remote_path)
-
-    if (isBeacon):
-        cd_results = await cd_results
-    
-    return f"{cd_results}"
-
-async def execute(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    # TODO: get these from function parameters and extract in the parent function instead
-    exe = taskData.args.get_arg('exe')
-    args = taskData.args.get_arg('args')
-    output = taskData.args.get_arg('output')
-
-    execute_results = await interact.execute(exe=exe, args=args, output=output)
-
-    if (isBeacon):
-        execute_results = await execute_results
-
-    return execute_results
-
-async def mkdir(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    # TODO: get these from function parameters and extract in the parent function instead
-    remote_path = taskData.args.get_arg('path')
-
-    mkdir_results = await interact.mkdir(remote_path=remote_path)
-
-    if (isBeacon):
-        mkdir_results = await mkdir_results
-
-    return mkdir_results
-
-async def pwd(taskData: PTTaskMessageAllData):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    pwd_results = await interact.pwd()
-
-    if (isBeacon):
-        pwd_results = await pwd_results
-
-    return f"{pwd_results}"
-
-async def terminate(taskData: PTTaskMessageAllData, pid: int):
-    interact, isBeacon = await create_sliver_interact(taskData)
-
-    terminate_results = await interact.terminate(pid=pid)
-
-    if (isBeacon):
-        terminate_results = await terminate_results
-
-    return f"{terminate_results}"
 
 async def shell(taskData: PTTaskMessageAllData):
     interact, isBeacon = await create_sliver_interact(taskData)
