+++ Payload_Type/sliverimplant/sliverimplant/agent_functions/shell.py
@@ -50,3 +50,4 @@ class Shell(CommandBase):
     async def process_response(self, task: PTTaskMessageAllData, response: any) -> PTTaskProcessResponseMessageResponse:
         resp = PTTaskProcessResponseMessageResponse(TaskID=task.Task.ID, Success=True)
         return resp
+
